Update seed scripts to write headers to CSV files
